requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.3.0
fastapi==0.103.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
joblib==1.3.2
mysqlclient==1.4.6
numpy==1.21.6
packaging==24.0
pandas==1.1.5
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.31.0
schedule==1.2.2
scikit-learn==0.22.2.post1
scipy==1.7.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.4.1
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.15
uvicorn==0.22.0
whitenoise==6.5.0
xlwt==1.3.0
zipp==3.15.0
